+++ src/FixedSource.cpp
@@ -18,41 +18,45 @@
 FixedSource::FixedSource(const pugi::xml_node& root)
     : Driver{root}, source{root.child("problemtype").child("fixedsource")} {};
 
-Estimator FixedSource::Solve() {
-  std::vector<std::future<Estimator>> results;
+EstimatorSet FixedSource::Solve() {
+  std::vector<std::future<EstimatorSet>> worker_estimator_sets;
   std::cout << "Spawning " << std::to_string(threads) << " threads working on "
             << std::to_string(batchsize) << " histories..." << std::endl;
   for (size_t i = 0; i < threads; i++) {
-    results.push_back(std::async(&FixedSource::StartWorker, this));
+    worker_estimator_sets.push_back(
+        std::async(&FixedSource::StartWorker, this));
   }
-  return std::reduce(
-             results.begin(), results.end(), Estimator{},
-             [](auto& accumulated, auto& future) {
-               return accumulated += future.get();
-             })
-      .Normalize(batchsize);
+  EstimatorSet solver_estimator_set = init_estimator_set;
+  for (auto& worker_estimator_set : worker_estimator_sets) {
+    solver_estimator_set += worker_estimator_set.get();
+  }
+  return solver_estimator_set.Normalize(batchsize);
 }
 
 //// private
 
-Estimator FixedSource::StartWorker() {
-  Estimator worker_estimator;
+EstimatorSet FixedSource::StartWorker() {
+  // we _could_ return a newly constructed EstimatorSet from each call to
+  // Transport() to keep it purely functional, but I suspect it will be super
+  // slow, so we construct it once at the beginning of a thread and pass it
+  // as a reference parameter to Transport()
+  EstimatorSet worker_estimator_set = init_estimator_set;
   while (true) {
     // atomically update thread-local count of histories started
     auto elapsed = histories_elapsed++;
     if (elapsed >= batchsize) {
       break;
     }
+    std::list<Particle> bank;
     // Use the remaining number of histories run for the seed
-    auto p{source.Sample(seed + elapsed)};
-    // we choose a list because list::splice is constant time
-    std::list<Particle> bank(1, p);
+    bank.emplace_back(source.Sample(seed + elapsed));
     while (!bank.empty()) {
-      auto result = transport_method->Transport(bank.back(), world);
+      transport_method->Transport(bank.back(), worker_estimator_set, world);
+      // new Particle objects are added to front
+      bank.back().MoveSecondariesToFrontOf(bank);
+      // most recently processed Particle is removed from back
       bank.pop_back();
-      worker_estimator += result.estimator;
-      bank.splice(bank.begin(), result.banked);
     }
   }
-  return worker_estimator;
+  return worker_estimator_set;
 }