+++ src/FixedSource.hpp
@@ -14,11 +14,11 @@ public:
   /// @brief Creates objects necessary for a fixed source calculation
   FixedSource(const pugi::xml_node& root);
   /// @brief Spawn workers to work on chunks of history
-  Estimator Solve() override;
+  EstimatorSet Solve() override;
 
 private:
   // function executed by a worker on a single thread
-  Estimator StartWorker();
+  EstimatorSet StartWorker();
   // fixed source from which new Particle objects can be sampled from
   const Source source;
   // Number of histories completed or initiated by all threads. May exceed