+++ src/TransportMethod.cpp
@@ -5,6 +5,7 @@
 #include "Estimator.hpp"
 #include "Material.hpp"
 #include "Nuclide.hpp"
+#include "Particle.hpp"
 #include "Reaction.hpp"
 #include "World.hpp"
 
@@ -12,17 +13,6 @@
 #include <random>
 #include <stdexcept>
 #include <string>
-
-// TransportMethod::Outcome
-
-//// public
-
-TransportMethod::Outcome&
-TransportMethod::Outcome::operator+=(Outcome&& rhs) noexcept {
-  estimator += rhs.estimator;
-  banked.splice(banked.begin(), rhs.banked);
-  return *this;
-}
 
 // TransportMethod
 
@@ -57,48 +47,42 @@ TransportMethod::~TransportMethod() noexcept {}
 
 // SurfaceTracking
 
-// SurfaceTracking
-
-TransportMethod::Outcome
-SurfaceTracking::Transport(Particle& p, const World& w) const noexcept {
-  Outcome result;
+Bank SurfaceTracking::Transport(
+    Particle& p, EstimatorSet& e, const World& w) const noexcept {
+  Bank history_bank;
   p.SetCell(w.FindCellContaining(p.GetPosition()));
-  while (p.alive) {
+  while (p.IsAlive()) {
     const auto distance_to_collision = std::exponential_distribution{
         p.GetCell().material->number_density *
         p.GetCell().material->GetMicroscopicTotal(p)}(p.rng);
     const auto [nearest_surface, distance_to_surface_crossing] =
         p.GetCell().NearestSurface(p.GetPosition(), p.GetDirection());
     if (distance_to_collision < distance_to_surface_crossing) {
-      result.estimator.at(Estimator::Event::collision) += 1;
       p.Stream(distance_to_collision);
       const auto& nuclide = p.SampleNuclide();
-      result.estimator.at(Estimator::Event::implicit_fission) +=
-          nuclide.GetNuBar(p) * nuclide.GetReaction(p, Reaction::fission) /
-          nuclide.GetTotal(p);
       nuclide.Interact(p);
     }
     else {
-      result.estimator.at(Estimator::Event::surface_crossing) += 1;
       p.Stream(distance_to_surface_crossing + constants::nudge);
       p.SetCell(w.FindCellContaining(p.GetPosition()));
-      if (!p.GetCell().material) {
-        p.Kill();
-      }
+      p.current_surface = nearest_surface;
+      p.event = p.GetCell().material ? Particle::Event::surface_cross
+                                     : Particle::Event::leak;
     }
+    e.Score(p);
   }
-  return result;
+  return history_bank;
 }
 
 // CellDeltaTracking
 
 // TODO: Remove embarrasing amount of code duplication between this and
 // SurfaceTracking
-TransportMethod::Outcome
-CellDeltaTracking::Transport(Particle& p, const World& w) const noexcept {
-  Outcome result;
+Bank CellDeltaTracking::Transport(
+    Particle& p, EstimatorSet& e, const World& w) const noexcept {
+  Bank history_bank;
   p.SetCell(w.FindCellContaining(p.GetPosition()));
-  while (p.alive) {
+  while (p.IsAlive()) {
     const auto distance_to_collision = std::exponential_distribution{
         p.GetCell().material->number_density *
         p.GetCell().material->GetMicroscopicMajorant(p)}(p.rng);
@@ -107,28 +91,25 @@ CellDeltaTracking::Transport(Particle& p, const World& w) const noexcept {
     if (distance_to_surface_crossing < distance_to_collision) {
       // a surface crossing occurs
       p.Stream(distance_to_surface_crossing + constants::nudge);
-      result.estimator.at(Estimator::Event::surface_crossing) += 1;
       p.SetCell(w.FindCellContaining(p.GetPosition()));
-      if (!p.GetCell().material){
-        p.Kill();
-      }
+      p.current_surface = nearest_surface;
+      p.event = p.GetCell().material ? Particle::Event::surface_cross
+                                     : Particle::Event::leak;
     }
     else if (std::bernoulli_distribution{
                  p.GetCell().material->GetMicroscopicTotal(p) /
                  p.GetCell().material->GetMicroscopicMajorant(p)}(p.rng)) {
       // a real collision occurs
       p.Stream(distance_to_collision);
-      result.estimator.at(Estimator::Event::collision) += 1;
       const auto& nuclide = p.SampleNuclide();
-      result.estimator.at(Estimator::Event::implicit_fission) +=
-          nuclide.GetNuBar(p) * nuclide.GetReaction(p, Reaction::fission) /
-          nuclide.GetTotal(p);
       nuclide.Interact(p);
     }
     else {
       // a virtual collision occurs, do nothing
       p.Stream(distance_to_collision);
+      p.event = Particle::Event::virtual_collision;
     }
+    e.Score(p);
   }
-  return result;
+  return history_bank;
 }