+++ src/World.hpp
@@ -28,6 +28,10 @@ public:
   ///        dependence
   /// @todo Check each Cell if it has constant temperature dependence
   bool HasConstantTemperature() const noexcept;
+  /// @brief Returns the CSGSurface with the given name
+  /// @exception std::runtime_error CSGSurface with given name not found
+  std::shared_ptr<const CSGSurface>
+  FindSurfaceByName(const std::string& name) const;
 
 private:
   // Helper function to create all CSGSurface objects that appear in Cell